+++ Coimbra/ActorSceneManagerAPI.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using UnityEditor;
+using UnityEngine;
 using UnityEngine.SceneManagement;
 
 namespace Coimbra
@@ -6,8 +7,15 @@ namespace Coimbra
     /// <summary>
     /// Custom <see cref="SceneManagerAPI"/> that does some pre-processing when unloading a scene.
     /// </summary>
+    [InitializeOnLoad]
     public class ActorSceneManagerAPI : SceneManagerAPI
     {
+        static ActorSceneManagerAPI()
+        {
+            EditorApplication.playModeStateChanged -= HandlePlayModeStateChanged;
+            EditorApplication.playModeStateChanged += HandlePlayModeStateChanged;
+        }
+
         /// <summary>
         /// Call this method from your custom <see cref="SceneManagerAPI"/> if not using the <see cref="ActorSceneManagerAPI"/>.
         /// </summary>
@@ -62,6 +70,20 @@ namespace Coimbra
         private static void HandleBeforeSceneLoad()
         {
             overrideAPI ??= new ActorSceneManagerAPI();
+        }
+
+        private static void HandlePlayModeStateChanged(PlayModeStateChange playModeStateChange)
+        {
+            switch (playModeStateChange)
+            {
+                case PlayModeStateChange.EnteredEditMode:
+                case PlayModeStateChange.ExitingEditMode:
+                {
+                    Actor.OnPlayModeStateChanged();
+
+                    break;
+                }
+            }
         }
     }
 }